backend/feedback_system/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
psycopg2==2.9.10
PyJWT==2.10.1
regex==2024.11.6
requests==2.32.3
sqlparse==0.5.3
textblob==0.18.0.post0
tqdm==4.67.1
urllib3==2.3.0

